Add `exec.clean` parameter to template
We need this so as not to confuse tmux, at least for desktop users.
The problem is very rare, but it can be reproduced when using tmux
on the host and you start/restart the jail (within the tmux session),
so the vm-bhyve rc(8) script will try to start the virtual machine
with the user's environment, confusing tmux, resulting in the error
"error creating /tmp/tmux-<host user UID>/default (No such file or directory)"
(which is not easily visible) and, worse, the virtual machine will
not start unless it is started manually.

Another alternative is the following patch:
 https://github.com/freebsd/vm-bhyve/pull/25